build: support 0.35.0

+++ src/placeos-driver/protocol/management.cr
@@ -311,9 +311,9 @@ class PlaceOS::Driver::Protocol::Management
     @launch_count += 1
     @launch_time = Time.utc.to_unix
 
-    fetch_pid = Promise.new(Int32)
+    fetch_pid = Promise.new(Int64)
     spawn(same_thread: true) { launch_driver(fetch_pid, stdin_reader, stderr_writer) }
-    @pid = fetch_pid.get.as(Int32)
+    @pid = fetch_pid.get.as(Int64)
 
     # Start processing the output of the driver
     loaded = Promise.new(Nil)
